SHA-1: 357f6b7027256db8a38f5d27b606f267a47a703e
* For running on Windows10 build 18363.592 with Docker Desktop 2.2.0.3(42716). Updated Docker config files and installer wrappers to update the clean_docker/magento2 project to working with as of today's modern levels of PHP, MYSQL8, Apache, and the base Debian. Project here will install and seems to work ok from a quick web run through. Have not run unit tests.

To get MYSQL8 to work, you need to add a custom cnf file to your docker root. Create path C:\Users\<username>\.docker\mysqlconf and create my.cnf with these lines:

[mysqld]
default-authentication-plugin=mysql_native_password
log_bin_trust_function_creators=1

use of LF EOL unix style may be important.

Note that this dir has been mapped in the docker-compose.yml for db service.

volumes:
      - ~/.docker/mysqlconf:/etc/mysql/conf.d

With out this changes the installer will not be able to connect to install the db or complete.

Remember to create the mysql dir for the db volumn mount. Also give Full windows to permissions to the src and the mysql dir as that seems to make the mount sync work better.